+++ socket/handlers/matching/matchingSocketListeners.js
@@ -1,234 +1,7 @@
-const { fetchMatchingApi, matchingFoundApi, matchingSuccessApi, updateMatchingStatusApi, updateBothMatchingStatusApi } = require("../../apis/matchApi");
-const { getSocketIdByMatchingUuid } = require("../../common/memberSocketMapper");
-const { updateOtherPriorityTrees, updatePriorityTree, handleSocketError, joinGameModeRoom, findMatching } = require("./matchingHandler/matchingStartedHandler");
-const { isSocketActiveAndInRoom } = require("./matchingHandler/matchingCommonHandler");
-const { deleteMySocketFromMatching } = require("./matchingHandler/matchingFoundHandler");
-const { emitError } = require("../../emitters/errorEmitter");
-const log = require("../../../common/customLogger");
+const { handleMatchingRequest, handleMatchingRetry } = require("./matchingHandler/matchingRequestHandler");
+const { handleMatchingFoundSuccess, handleMatchingSuccessReceiver, handleMatchingSuccessFinal } = require("./matchingHandler/matchingSuccessHandler");
+const {handleMatchingReject, handleMatchingNotFound, handleMatchingFail, handleMatchingQuit} = require("./matchingHandler/matchingFailHandler");
 
-const {
-  emitMatchingStarted,
-  emitMatchingFoundReceiver,
-  emitMatchingFoundSender,
-  emitMatchingSuccessSender,
-  emitMatchingSuccess,
-  emitMatchingFail,
-} = require("../../emitters/matchingEmitter");
-
-/**
- * "matching-request" - 매칭 요청 핸들러
- */
-async function handleMatchingRequest(socket, io, request) {
-  const threshold = request.threshold;
-  const gameMode = request.gameMode;
-  socket.data.matching.gameMode = gameMode;
-  const roomName = "GAMEMODE_" + gameMode;
-  socket.data.matching.roomName = roomName;
-
-  log.info("matching-request", socket);
-
-  // 2) socket.id가 소켓 룸 "GAMEMODE_" + gameMode에 있는지 확인
-  const usersInRoom = io.sockets.adapter.rooms.get(roomName) || new Set();
-  if (usersInRoom.has(socket.id)) {
-    emitError(socket, "You are already in the matching room for this game mode.");
-    return;
-  }
-
-  // 3) 게임 모드에 따라 room에 join
-  joinGameModeRoom(socket, io, roomName);
-
-  try {
-    // 4) 8080서버에 우선순위 계산 API 요청
-    const result = await fetchMatchingApi(socket, request);
-
-    // 6) API 정상 응답 받음
-    if (result) {
-      socket.data.matching.myMatchingInfo = result.myMatchingInfo;
-      socket.data.matching.matchingUuid = result.myMatchingInfo.matchingUuid;
-
-      // 7) "matching-started"emit
-      emitMatchingStarted(socket, result.myMatchingInfo);
-
-      log.debug(`myPriorityList : ${result.myPriorityList}`,socket);
-
-      // 8) 내 우선순위 트리 갱신
-      updatePriorityTree(socket, result.myPriorityList);
-
-      // 9) room에 있는 모든 socket의 우선순위 트리 갱신
-      await updateOtherPriorityTrees(io, socket, result.otherPriorityList);
-    }
-
-    // 10) priorityTree의 maxNode가 기준 점수를 넘는지 확인
-    const receiverSocket = await findMatching(socket, io, threshold);
-
-    if (receiverSocket && receiverSocket!=socket) {
-      // 11) receiverSocket이 매칭 room에 존재하는지 여부 확인
-      log.debug(`#11 check receiverSocket is in matching room, receiverSocket's memberId : ${receiverSocket.memberId}`, socket);
-      isSocketActiveAndInRoom(receiverSocket, io, roomName);
-
-      const matchingFoundReceiverRequest = {};
-
-      // 원하는 값들을 각각 프로퍼티에 할당
-      matchingFoundReceiverRequest.senderMatchingInfo = socket.data.matching.myMatchingInfo;
-      matchingFoundReceiverRequest.receiverMatchingUuid = receiverSocket.data.matching.matchingUuid;
-      
-      // 12) "matching-found-receiver" emit
-      emitMatchingFoundReceiver(receiverSocket, matchingFoundReceiverRequest);
-    }
-
-  } catch (error) {
-    console.log(error);
-    handleSocketError(socket, error);
-  }
-}
-
-async function handleMatchingRetry(socket, io, request) {
-
-}
-
-/**
- * 매칭 성공 요청 핸들러 (receiver가 보냄)
- */
-async function handleMatchingFoundSuccess(socket, io, request) {
-  log.info(`matching found success`, socket);
-
-  const matchingUuid=socket.data.matching.matchingUuid;
-  const senderMatchingUuid=request.senderMatchingUuid;
-  const senderSocket = await getSocketIdByMatchingUuid(io, senderMatchingUuid);
-
-  console.log("matching found success");
-  console.log(matchingUuid);
-  console.log(senderMatchingUuid);
-  console.log('sender : ',senderSocket.memberId);
-  console.log('receiver : ',socket.memberId);
-
-  if (!senderSocket) {
-    log.error(`Sender socket not found for senderMatchingUuid:${request.senderMatchingUuid}`, socket);
-    return;
-  }
-
-  // 15. 두 소켓의 matchingTargetUuid 값 저장
-  socket.data.matching.matchingTargetUuid = senderMatchingUuid;
-  senderSocket.data.matching.matchingTargetUuid = matchingUuid;
-  
-  // 16 ~ 18 데이터 삭제
-  const roomName = "GAMEMODE_" + socket.data.matching.gameMode;
-  deleteMySocketFromMatching(socket, io, roomName);
-  deleteMySocketFromMatching(senderSocket, io, roomName);
-
-  try {
-    const result = await matchingFoundApi(socket, matchingUuid, senderMatchingUuid);
-    if (result) {
-      emitMatchingFoundSender(senderSocket, result.myMatchingInfo);
-    }
-  } catch (error) {
-    log.error(`matching-found-success : ${error.message}`, socket);
-    handleSocketError(socket, error);
-  }
-}
-
-//TODO: 
-// 25. "matching-success-receiver"
-async function handleMatchingSuccessReceiver(socket,io,request){
-  console.log(request);
-  const senderSocket = await getSocketIdByMatchingUuid(io,request.senderMatchingUuid);
-  // 26. "matching-success-sender" emit
-  emitMatchingSuccessSender(senderSocket);
-}
-
-/**
- * 매칭 성공 최종 단계 핸들러 (sender가 보냄)
- */
-async function handleMatchingSuccessFinal(socket, io) {
-  log.info(`Received 'matching-success-final' from socketId:${socket.id}`, socket);
-
-  const receiverSocket = await getSocketIdByMatchingUuid(io, socket.data.matching.matchingTargetUuid);
-  if (!receiverSocket) {
-    log.warn(`Receiver socket not found for matchingTargetUuid:${socket.data.matching.matchingTargetUuid}`, socket);
-    return;
-  }
-
-  try {
-    const result = await matchingSuccessApi(socket);
-    console.log(result);
-    if (result) {
-      socket.join("CHAT_" + result);
-      receiverSocket.join("CHAT_" + result);
-      emitMatchingSuccess(socket, receiverSocket, result);
-    }
-  } catch (error) {
-    log.error(`matching-success-final : ${error.message}`, socket);
-    handleSocketError(socket, error);
-  }
-}
-
-/**
- * 매칭 거절 핸들러
- */
-async function handleMatchingReject(socket, io) {
-  log.info(`Received 'matching-reject' from socketId:${socket.id}`);
-
-  const otherSocket = await getSocketIdByMatchingUuid(io, socket.data.matching.matchingTargetUuid);
-  if (socket.data.matching.gameMode) {
-    try {
-      await updateBothMatchingStatusApi(socket, "FAIL", socket.data.matching.matchingTargetUuid);
-    } catch (error) {
-      log.error(`matching-reject : ${error.message}`, socket);
-      handleSocketError(socket, error);
-    }
-  }
-
-  if (otherSocket) {
-    emitMatchingFail(otherSocket);
-    otherSocket.data.matching.matchingTargetUuid = null;
-  }
-
-  emitMatchingFail(socket);
-  socket.data.matching.matchingTargetUuid = null;
-}
-
-async function handleMatchingNotFound(socket){
-
-}
-
-/**
- * 매칭 실패 핸들러
- */
-async function handleMatchingFail(socket) {
-  log.info(`matching fail`, socket);
-
-  if (socket.data.matching.gameMode) {
-    try {
-      await updateMatchingStatusApi(socket, "FAIL");
-    } catch (error) {
-      log.error(`Error in 'matching-fail' for socketId:${socket.id}, error: ${error.message}`);
-      handleSocketError(socket, error);
-      return;
-    }
-  }
-
-  socket.data.matching.matchingTargetUuid = null;
-  emitMatchingFail(socket);
-}
-
-/**
- * 매칭 종료 핸들러
- */
-async function handleMatchingQuit(socket, io) {
-  log.info(`matching-quit`, socket);
-
-  if (socket.data.matching.gameMode) {
-    try {
-      await updateMatchingStatusApi(socket, "QUIT");
-    } catch (error) {
-      log.error(`matching-quit : ${error.message}`, socket);
-      handleSocketError(socket, error);
-      return;
-    }
-  }
-
-  deleteMySocketFromMatching(socket, io, socket.data.matching.roomName);
-}
 
 /**
  * Socket 이벤트 리스너 등록
@@ -237,7 +10,7 @@ function setupMatchSocketListeners(socket, io) {
   socket.on("matching-request", (request) => handleMatchingRequest(socket, io, request));
   socket.on("matching-retry", (request) => handleMatchingRetry(socket, io, request));
   socket.on("matching-found-success", (request) => handleMatchingFoundSuccess(socket, io, request));
-  socket.on("matching-success-receiver", (request) => handleMatchingSuccessReceiver(socket, io, request));
+  socket.on("matching-success-receiver", (request) => handleMatchingSuccessReceiver(io, request));
   socket.on("matching-success-final", () => handleMatchingSuccessFinal(socket, io));
   socket.on("matching-not-found", () => handleMatchingNotFound(socket, io));
   socket.on("matching-reject", () => handleMatchingReject(socket, io));
